requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.10.0
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-leaflet==0.32.0
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
google-auth==2.42.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
kombu==5.5.4
oauthlib==3.3.1
onesignal==0.1.3
onesignal-sdk==2.0.0
packaging==25.0
phonenumbers==9.0.17
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pystache==0.6.8
pytest==8.4.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
static==1.1.1
typing==3.7.4.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
